Initial commit
This is the first commit for the Odin Project Connect Four TDD Project.
The purpose of this project is to practice building a project using
TDD principles. Currently tests are working for the initial game setup
and the switching of the current active player. Creation, display, and
updating of the game grid are currently functional.