Added functionality for game_loop, other methods
With the added functionality of the main game_loop, the Connect Four
game is now operational and playable in the command line.
Inside of board.rb, the #all_spaces_full? method was added to check
for a full game board. If the board is full, the game_loop performs one
final check for a winner before declaring a draw and exiting the program.

Final steps will be to go back through and clean up the code, removing
any deprecated/unused code, as well as adding in appopriate comments.

This project began as a rather frustrating exercise in TDD, but has been
very enlightening in demonstrating the value of this approach. Once the
problem was broken down in appropriately small pieces suitable for
testing, the overall shape of the code and needed functions became much
clearer than it would have been working from scratch with no testing
framework to guide the project design. While this approach seems to be
more labor and thought intensive at the outset, I believe the resulting
program was much more streamlined and elegant in the end. Much success.