Cleaned up redundant methods, added comments
In the original board.rb, update_board, choose_column, and current_row
were all handling the selection of any empty row when placing a piece.
This required the use of an instance variable @available_rows, which kept
track of each columns first empty row. This functionality has been moved
to update_board, which starts at the bottom of the column and works its
way up until it finds an empty row. This may be causing more processing
usage on each pass, but at the same time, is not keeping track of an
additional instance variable and cuts down on other method calls.

Tests were updated to reflect these changes. Main_spec.rb has broken due
to something with the game_loop accepting user input. Maybe commenting
out this section for testing would work? I am leaving as is, since the
majority of tests are located in board_spec, which is working as
intended.